% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_content}
\alias{algolia_content}
\title{Get Algolia "content"}
\usage{
algolia_content(config = create_config())
}
\arguments{
\item{config}{config as retrieved by create_config() (default:
create_config())}
}
\value{
data.frame with content of Algolia index
}
\description{
Get Algolia "content"
}
\examples{
\dontrun{
algolia_content()
}
}
