% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_delete_index}
\alias{algolia_delete_index}
\title{Delete Algolia Index}
\usage{
algolia_delete_index(config)
}
\arguments{
\item{config}{config as retrieved by create_config()}
}
\value{
deletes "index_name" (config$index_name) at Algolia URL (config$api_url)
in case that provided API Key has sufficient rights!
}
\description{
Delete Algolia Index
}
\examples{
\dontrun{
config <- create_config(app_id = "your_algolia_application_id",
                        api_key = "your_algolia_api_key",
                        index_name = "your_algolia_index_name",
                        path_index_json = "path_to_your_index.json")
algolia_content(config)
}
}
\seealso{
<https://www.algolia.com/doc/api-client/methods/api-keys/>
}
