% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_get}
\alias{algolia_get}
\title{Algolia "GET"}
\usage{
algolia_get(config)
}
\arguments{
\item{config}{config as retrieved by create_config()}
}
\description{
Algolia "GET"
}
\examples{
\dontrun{
config <- create_config(app_id = "your_algolia_application_id",
                        api_key = "your_algolia_api_key",
                        index_name = "your_algolia_index_name",
                        path_index_json = "path_to_your_index.json")
algolia_get(config)
}
}
