% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_post_batch}
\alias{algolia_post_batch}
\title{Performs Algolia Post Batch}
\usage{
algolia_post_batch(config)
}
\arguments{
\item{config}{as retrieved by create_config()}
}
\value{
performs batch operation at provided Algolia url (config$api_url)
in case of sufficient rights of the provided API key (see:
<https://www.algolia.com/doc/api-client/methods/api-keys/>)
}
\description{
Performs Algolia Post Batch
}
\examples{
\dontrun{
config <- create_config(app_id = "your_algolia_application_id",
                        api_key = "your_algolia_api_key",
                        index_name = "your_algolia_index_name",
                        path_index_json = "path_to_your_index.json")
algolia_post_batch(config)
}
}
\seealso{
<https://www.algolia.com/doc/rest-api/search/#batch-write-operations>
}
