% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{check_if_path_or_url}
\alias{check_if_path_or_url}
\title{Helper Function: Check If Path or URL}
\usage{
check_if_path_or_url(path)
}
\arguments{
\item{path}{path/url to be checked}
}
\value{
TRUE (if existing) or FALSE (if not existing)
}
\description{
Helper Function: Check If Path or URL
}
\examples{
\dontrun{
check_if_path_or_url(path = "public/index.json")
}

}
