% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_threshold_kb.R
\name{check_threshold_kb}
\alias{check_threshold_kb}
\title{Check Algolia Threshold KB}
\usage{
check_threshold_kb(index_list, threshold_kb = 10)
}
\arguments{
\item{index_list}{as retrieved by algoliar::read_index_list()}

\item{threshold_kb}{(default: 10), 10 KB is threshold for free Algolia usage,
20 KB for paid for details: https://www.algolia.com/doc/faq/basics/is-there-a-size-limit-for-my-index-records/}
}
\value{
index_list respecting Algolia threshold KB, by reducing the "content"
length accordingly
}
\description{
Check Algolia Threshold KB
}
