% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{create_addObject}
\alias{create_addObject}
\title{Create addObject list}
\usage{
create_addObject(index_list)
}
\arguments{
\item{index_list}{as retrieved by read_index_list()}
}
\value{
list as required for API endpoint "addObject"
}
\description{
prepare json for API endpoint "addObject"
}
\examples{
\dontrun{
index_list <- read_index_list(path = "public/index.json")
create_addObject(index_list)
}

}
\seealso{
<https://www.algolia.com/doc/api-reference/api-methods/add-objects/>
}
