% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{create_config}
\alias{create_config}
\title{Create Algolia Configuration}
\usage{
create_config(app_id = Sys.getenv("algolia_app_id"),
  api_key = Sys.getenv("algolia_api_key"),
  index_name = Sys.getenv("algolia_index_name"),
  path_index_json = "public/index.json")
}
\arguments{
\item{app_id}{app_id Application ID. This is your unique application identifier.
It's used to identify you when using Algolia's API. (see: create a API key
with valid scope in the Algolia Dashboard), (default: Sys.getenv("algolia_app_id"))}

\item{api_key}{api_key (default: Sys.getenv("algolia_api_key"),
<https://www.algolia.com/doc/api-client/methods/api-keys/>}

\item{index_name}{name of Algolia index name (default:
Sys.getenv("algolia_index_name"))}

\item{path_index_json}{path to index.json (default: "public/index.json")}
}
\value{
list with algolia config
}
\description{
Create Algolia Configuration
}
\examples{
\dontrun{
config <- create_config(app_id = "your_algolia_application_id",
                        api_key = "your_algolia_api_key",
                        index_name = "your_algolia_index_name",
                        path_index_json = "path_to_your_index.json")
}
}
\seealso{
<https://www.algolia.com/doc/guides/getting-started/quick-start/tutorials/getting-started-with-the-dashboard/>
}
