% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{read_index_json}
\alias{read_index_json}
\title{Read Index.Json}
\usage{
read_index_json(path = "public/index.json")
}
\arguments{
\item{path}{path or url to index.json (default: "public/index.json")}
}
\value{
imported index.json
}
\description{
Read Index.Json
}
\examples{
\dontrun{
index_json <- read_index_json(path = "public/index.json")
}

}
