% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{read_index_list}
\alias{read_index_list}
\title{Read Index.Json As List}
\usage{
read_index_list(path = "public/index.json", encoding = "UTF-8", ...)
}
\arguments{
\item{path}{path or url to index.json (default: "public/index.json")}

\item{encoding}{encoding (default: "UTF-8")}

\item{...}{additional arguments passed to jsonlite::read_json()}
}
\value{
imported index.json as list
}
\description{
Read Index.Json As List
}
\examples{
\dontrun{
index_list <- read_index_list(path = "public/index.json")
}

}
