% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_content}
\alias{algolia_content}
\title{Get Algolia "content"}
\usage{
algolia_content(config)
}
\arguments{
\item{config}{config as retrieved by create_config()}
}
\value{
data.frame with content of Algolia index
}
\description{
Get Algolia "content"
}
\examples{
\dontrun{
config <- create_config(app_id = "your_algolia_application_id",
                        api_key = "your_algolia_api_key",
                        index_name = "your_algolia_index_name",
                        path_index_json = "path_to_your_index.json")
algolia_content(config)
}
}
