% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_delete_index}
\alias{algolia_delete_index}
\title{Delete Algolia Index}
\usage{
algolia_delete_index(config = create_config())
}
\arguments{
\item{config}{config as retrieved by create_config() (default:
create_config())}
}
\value{
deletes "index_name" (config$index_name) at Algolia URL (config$api_url)
in case that provided API Key has sufficient rights!
}
\description{
Delete Algolia Index
}
\examples{
\dontrun{
algolia_content()
}
}
\seealso{
<https://www.algolia.com/doc/api-client/methods/api-keys/>
}
