% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_get}
\alias{algolia_get}
\title{Algolia "GET"}
\usage{
algolia_get(config = create_config())
}
\arguments{
\item{config}{config as retrieved by create_config() (default:
create_config())}
}
\description{
Algolia "GET"
}
\examples{
\dontrun{
algolia_get()
}
}
