% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{algolia_post_batch}
\alias{algolia_post_batch}
\title{Performs Algolia Post Batch}
\usage{
algolia_post_batch(config = create_config(), threshold_kb = 10)
}
\arguments{
\item{config}{config as retrieved by create_config() (default:
create_config())}

\item{threshold_kb}{(default: 10), 10 KB is threshold for free Algolia usage,
20 KB for paid for details: https://www.algolia.com/doc/faq/basics/is-there-a-size-limit-for-my-index-records/}
}
\value{
performs batch operation at provided Algolia url (config$api_url)
in case of sufficient rights of the provided API key (see:
<https://www.algolia.com/doc/api-client/methods/api-keys/>)
}
\description{
Performs Algolia Post Batch
}
\examples{
\dontrun{
algolia_post_batch()
}
}
\seealso{
<https://www.algolia.com/doc/rest-api/search/#batch-write-operations>
}
