% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algolia.R
\name{prepare_batch_json}
\alias{prepare_batch_json}
\title{Pepare Batch Json}
\usage{
prepare_batch_json(config, encoding = "UTF-8", ...)
}
\arguments{
\item{config}{config as retrieved by create_config()}

\item{encoding}{encoding (default: "UTF-8")}

\item{...}{additional arguments passed to jsonlite::read_json()}
}
\description{
Pepare Batch Json
}
\examples{
\dontrun{
config <- create_config(app_id = "your_algolia_application_id",
                        api_key = "your_algolia_api_key",
                        index_name = "your_algolia_index_name",
                        path_index_json = "path_to_your_index.json")
batch_json <- prepare_batch_json(config)
}
}
